Changed Performance Data Output
Changed the performance data output to follow the guide lines https://nagios-plugins.org/doc/guidelines.html#AEN200
Removed the C from the temperature output, as C is not allowed.
Added remaining_time_warn to the perfdata output and converted to the time to seconds in the perdata output to follow the guidelines. This will help tools like rrd to display the graphs.